+++ CHANGELOG.md
@@ -21,6 +21,10 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+### New
+
+* improved performance when repeating conversions between color types
+
 ### B/C Breaks
 
 * component values (`.red()` et al) are no longer available everywhere